server/requirements.txt
fastapi
pydantic
requests
PyPDF2
openai
python-dotenv
httpx
uvicorn